Add account activation page
After registering, users will receive a mail pointing them to this page
with the user ID and a token associated as query parameters. Users
cannot login before activating their account. When accessing the account
activation page, the frontend will pass the request over to the backend,
which will then activate the account if the token is valid.